backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
sqlalchemy==2.0.23
slowapi==0.1.9
python-multipart==0.0.6
